third_party/WebKit/LayoutTests/bluetooth/characteristic/getDescriptor/descriptor-not-found.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
   return setBluetoothFakeAdapter('DisconnectingHealthThermometerAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['health_thermometer']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('health_thermometer'))
    .then(service => service.getCharacteristic('measurement_interval'))
    .then(characteristic => {
      assert_promise_rejects_with_message(
        // Valid Range is not present in this characteristic
        characteristic.getDescriptor(2906),
        new DOMException(
          'No Descriptors with specified UUID found in Characteristic.',
          'NotFoundError'));
    })
}, 'Request for absent descriptor. Reject with NotFoundError.');
</script>
